better typing of Distribution objects
For exampler, rand(Distribution(Pair, Int, Int), Dict, 10)
was creating a Dict{Any,Any}. Also, rand(1:10, IntSet)
was not working because IntSet is a concrete type.

Note: this is still imperfect, as
Distribution(Pair{Int}, 1:10, Float) will still deduce an eltype
of Pair{Int,Int} instead of Pair{Int,Float}, but it's not clear
if there is a use-case for that.